backend/src/requirements.txt
asgiref==3.2.10
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-rest-auth==1.1.1
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
djangorestframework==3.11.1
docutils==0.15.2
future==0.16.0
idna==2.7
jmespath==0.10.0
oauthlib==3.1.0
pathspec==0.5.9
Pillow==7.2.0
pyasn1==0.4.8
pypiwin32==223
python-dateutil==2.8.0
python3-openid==3.2.0
pytz==2020.1
pywin32==228
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.3
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
urllib3==1.24.3
wcwidth==0.1.9
